get, insert and update ratings user

+++ backend/models/movierating.js
@@ -3,7 +3,7 @@ module.exports = (sequelize, DataTypes) => {
   const MovieRating = sequelize.define('MovieRating', {
     UserId: DataTypes.INTEGER,
     MovieId: DataTypes.INTEGER,
-    rating: DataTypes.INTEGER
+    rating: DataTypes.DOUBLE
   }, {});
   MovieRating.associate = function(models) {
     // associations can be defined here
